% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{statscourse_skeleton}
\alias{statscourse_skeleton}
\title{Create statscourse project directory}
\usage{
statscourse_skeleton(
  path,
  assignment = "01",
  number = "45110000",
  yourname = "名 姓",
  thedate = "2014-04-01"
)
}
\arguments{
\item{path}{The path of the created project.}

\item{assignment}{The homework assignment number.}

\item{number}{The student id number.}

\item{yourname}{The student name.}

\item{thedate}{The date.}
}
\description{
This function is used to create the directory structure for the
basic statistics homework assignments. It will also automatically
create an RMarkdown file for the homework assignement.
}
